DALYs calculation

Calculate dalys

calculate_dalys(
  con,
  touchstone_name,
  year_min = 2000,
  year_max = 2030,
  vimc_dalys_only = TRUE,
  modelling_group = NULL,
  stochastic_data = NULL,
  dalys_parameters = NULL,
  life_table = NULL
)

create_dalys_parameters(con, touchstone_name = "201710gavi", vimc_dalys_only)

create_dalys_life_table(
  con,
  touchstone_name = "201710gavi",
  year_min = 2000,
  year_max = 2030
)

Arguments

con

You can be readonly user to run this function. But if you need to import dalys for Ferrari, Li and LiST (201710gavi), you can use import user.

touchstone_name

touchstone for which dalys are calcualted

year_min

minimal year

year_max

maximum year

vimc_dalys_only

set to be TRUE if we are only interested in Ferrari, Li and LiST (201710gavi)

modelling_group

This parameter makes the calcualtion more flexible, specify a vector of modelling_group(s) that you are interested in.

stochastic_data

If this is not NULL, then instead of querying the database for the burden estimate set, use the specified stochastic_data, which must be a data frame containing the columns "burden_estimate_set", "country", "year", "age", "burden_outcome" and "burden". The burden_estimate_set should refer to the central burden estimate set for that group, which daly parameters related to. "country" is the 3-character representation. "year" and "age" are trivial; "burden_outcome" is the integer code for the burden, for each line, and "burden" is the data value.

dalys_parameters

Leave as NULL, for single calls to calculate_dalys, but for many calls, for example stochastic runs, call create_dalys_parameters first, and pass the result as an argument here to speed things up.

life_table

Leave as NULL, for single calls to calculate_dalys, but for many calls, for example stochastic runs, call create_dalys_life_table first, and pass the result as an argument here to speed things up significantly.