Create a new touchstone

Create and import a new touchstone.

create_touchstone(
  con,
  dat,
  demography_from = NULL,
  path_meta = "meta",
  transaction = TRUE,
  dry_run = TRUE
)

Arguments

con

Database connection. You will need to be the vimc or import user (not readonly) to run this function.

dat

A data.frame of coverage data to import.

demography_from

Touchstone id to import demographic statistics from. This must currently be given, but in future we'll allow this to be imported from a csv

path_meta

Path that we look for various metadata files. Eventually we'll document what they look like.

transaction

Do in one transaction (logical scalar)

dry_run

Don't commit (just rollback the transaction after completion)