Impact Calculation (method 2)

Provide age-specific coverage-un_pop-fvps

fix_coverage_fvps(
  con,
  touchstone_name = "201710gavi",
  year_min = 2000,
  year_max = 2100,
  pine = FALSE,
  write_table = TRUE,
  report_suspecious_coverage = FALSE,
  touchstone_pop = NULL,
  gavi_support_levels = c("with", "bestminus")
)

Arguments

con

Database connection. You will need to be readonly user to run this function.

touchstone_name

Specify touchstone name only, not with specific version.

year_min

min year of vaccination

year_max

max year of vaccination

pine

this is for testthat. we only grab data for pine countries if true

write_table

If true, create a temporary table; otherwise return a dataframe

report_suspecious_coverage

switch on/off the reporting of suspecious coverage

touchstone_pop

population touchstone, this is for the modups where fvps are calculated using a coverage touchstone and a population touchstone

gavi_support_levels

specify gavi_support_levels that apply to the touchstone_name you are looking at